adds an option to control whether columns are inserted by default
This adds an `insert` option, analogous to the existing `select`
option, for use in the `Column` decorator.  Just like a column
specified with `@Column({select: false})` will not be selected by
default in queries, a column specified with `@Column({insert: false})`
will not be inserted by default in `INSERT`s.

This is the simplest intervention I could think of to allow for
propagating computed columns to typeorm entities.  Now, a column
can be specified as `@Column({insert: false, select: false})`
and if a query computes the column it will be preserved in
typeorm entities.  There are a lot of issues filed looking for similar
behavior, and it is on the roadmap, so at some point there may
be official typeorm support for something like this.